Migrate print and suggest commands
